porting 2022 oct 19 1438; started re-factoring of a_tables.rs; moved around stb_xxx modules.

+++ src/context.rs
@@ -579,7 +579,7 @@ pub struct ImGuiContext {
     pub CurrentTable: *mut ImGuiTable,
 
     // int                             TablesTempDataStacked;      // Temporary table data size (because we leave previous instances undestructed, we generally don't use TablesTempData.Size)
-    pub TablesTempDataStacked: c_int,
+    pub TablesTempDataStacked: usize,
 
     // ImVector<ImGuiTableTempData>    TablesTempData;             // Temporary table data (buffers reused/shared across instances, support nesting)
     pub TablesTempData: Vec<ImGuiTableTempData>,
